+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 ### Added
+- Introduced `dd()` to quickly dump list of passed arguments and halt script execution
 - Introduced `message_simple()` to show inline feedback message anywhere on the site
 - Password length is now actually enforced when saving options